refactored

+++ src/main/java/com/example/NachislenieZP/controllers/MainController.java
@@ -54,7 +54,7 @@ public final class MainController {
     }
 
     @PostMapping("/edit")
-    public String employeeEdit(@RequestParam int id, @RequestParam String name, @RequestParam String post, @RequestParam int salary, Model model) {
+    public String employeeEdit(@RequestParam long id, @RequestParam String name, @RequestParam String post, @RequestParam int salary, Model model) {
         EmployeeRepository.editEmployee(id, name, post, salary);
         return "redirect:/";
     }
